Split aggregate constructor checking
